+++ src/image.cpp
@@ -72,7 +72,7 @@ namespace Images
     {
         return ms_store.Find(name);
     }
-    slot Find(uint64_t hash)
+    slot Find(uint32_t hash)
     {
         return ms_store.Find(hash);
     }
