+++ src/image.h
@@ -14,5 +14,5 @@ namespace Images
     bool Exists(slot s);
     bool Exists(const char* name);
     slot Find(const char* name);
-    slot Find(uint64_t hash);
+    slot Find(uint32_t hash);
 };
