+++ src/pipeline.cpp
@@ -41,7 +41,7 @@ namespace Pipelines
     {
         return ms_store.Find(name);
     }
-    slot Find(uint64_t hash)
+    slot Find(uint32_t hash)
     {
         return ms_store.Find(hash);
     }
