+++ src/pipeline.h
@@ -10,5 +10,5 @@ namespace Pipelines
     const sg_pipeline* Get(slot s);
     bool Exists(slot s);
     slot Find(const char* name);
-    slot Find(uint64_t hash);
+    slot Find(uint32_t hash);
 };
