+++ src/shader.cpp
@@ -39,7 +39,7 @@ namespace Shaders
     {
         return ms_store.Find(name);
     }
-    slot Find(uint64_t hash)
+    slot Find(uint32_t hash)
     {
         return ms_store.Find(hash);
     }
